gitlab: Move webhook handler functions to dedicated files
The intention is to make handlers easier to find.
